chore: rename NIL_UUID -> NIL
