+++ src/index.js
@@ -2,8 +2,7 @@ export { default as v1 } from './v1.js';
 export { default as v3 } from './v3.js';
 export { default as v4 } from './v4.js';
 export { default as v5 } from './v5.js';
-export { default as NIL_UUID } from './nil_uuid.js';
-export { default as REGEX } from './regex.js';
+export { default as NIL } from './nil.js';
 export { default as version } from './version.js';
 export { default as validate } from './validate.js';
 export { default as stringify } from './stringify.js';
