+++ test/unit/version.test.js
@@ -1,10 +1,10 @@
 import assert from 'assert';
 import version from '../../src/version.js';
-import NIL_UUID from '../../src/nil_uuid.js';
+import NIL from '../../src/nil.js';
 
 describe('version', () => {
   test('check uuid version', () => {
-    assert.strictEqual(version(NIL_UUID), 0);
+    assert.strictEqual(version(NIL), 0);
 
     assert.strictEqual(version('d9428888-122b-11e1-b85c-61cd3cbb3210'), 1);
